Very incomplete and broken attempt to rip out JPA.
Just need to commit the WIP before what will be a very painful merge.

CBBF-53
