+++ bluebutton-server-app/pom.xml
@@ -35,41 +35,12 @@
 			<version>${hapi-fhir.version}</version>
 		</dependency>
 
-		<!-- This dependency includes the JPA server itself, which is packaged 
-			separately from the rest of HAPI FHIR -->
-		<dependency>
-			<groupId>ca.uhn.hapi.fhir</groupId>
-			<artifactId>hapi-fhir-jpaserver-base</artifactId>
-			<version>${hapi-fhir.version}</version>
-		</dependency>
-
 		<dependency>
 			<!-- Needed for JEE/Servlet support -->
 			<groupId>javax.servlet</groupId>
 			<artifactId>javax.servlet-api</artifactId>
 			<version>3.1.0</version>
 			<scope>provided</scope>
-		</dependency>
-
-		<dependency>
-			<!-- If you are using HAPI narrative generation, you will need to include 
-				Thymeleaf as well. Otherwise the following can be omitted. -->
-			<groupId>org.thymeleaf</groupId>
-			<artifactId>thymeleaf</artifactId>
-			<version>3.0.1.RELEASE</version>
-		</dependency>
-
-		<dependency>
-			<!-- Used for CORS support -->
-			<groupId>org.ebaysf.web</groupId>
-			<artifactId>cors-filter</artifactId>
-			<version>1.0.1</version>
-			<exclusions>
-				<exclusion>
-					<artifactId>servlet-api</artifactId>
-					<groupId>javax.servlet</groupId>
-				</exclusion>
-			</exclusions>
 		</dependency>
 
 		<dependency>
@@ -111,31 +82,10 @@
 		</dependency>
 
 		<dependency>
-			<!-- The Blue Button API runs against PostgreSQL in production. -->
-			<groupId>org.postgresql</groupId>
-			<artifactId>postgresql</artifactId>
-			<version>42.0.0</version>
-		</dependency>
-
-		<dependency>
-			<!-- The builds for this and other Blue Button API projects use HSQL in 
-				tests. -->
-			<groupId>org.hsqldb</groupId>
-			<artifactId>hsqldb</artifactId>
-			<version>2.2.4</version>
-		</dependency>
-
-		<dependency>
 			<!-- The test framework used for unit and integration tests. -->
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>com.phloc</groupId>
-			<artifactId>phloc-schematron</artifactId>
-			<version>2.7.1</version>
 		</dependency>
 	</dependencies>
 
@@ -368,7 +318,6 @@
 					</execution>
 				</executions>
 			</plugin>
-
 		</plugins>
 	</build>
