+++ bluebutton-server-app/src/main/java/gov/hhs/cms/bluebutton/server/app/BlueButtonStu3ServerConformanceProvider.java
@@ -1,0 +1,10 @@
+package gov.hhs.cms.bluebutton.server.app;
+
+import org.hl7.fhir.dstu3.hapi.rest.server.ServerConformanceProvider;
+
+/**
+ * 
+ */
+public final class BlueButtonStu3ServerConformanceProvider extends ServerConformanceProvider {
+
+}
