+++ bluebutton-server-app/src/main/java/gov/hhs/cms/bluebutton/server/app/stu3/rp/CapabilityStatementResourceProvider.java
@@ -1,0 +1,13 @@
+package gov.hhs.cms.bluebutton.server.app.stu3.rp;
+
+import org.hl7.fhir.dstu3.model.CapabilityStatement;
+
+import ca.uhn.fhir.rest.server.IResourceProvider;
+
+/**
+ * The HAPI {@link IResourceProvider} implementation for CCW STU3
+ * {@link CapabilityStatement}s.
+ */
+public final class CapabilityStatementResourceProvider implements IResourceProvider {
+
+}
