+++ bluebutton-server-app/src/main/java/gov/hhs/cms/bluebutton/server/app/stu3/rp/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Contains the HAPI {@link ca.uhn.fhir.rest.server.IResourceProvider}s for
+ * STU3.
+ */
+package gov.hhs.cms.bluebutton.server.app.stu3.rp;
